Reorganize classes
